[#27752] Display only valid AccountTypes for identity when editing System mapping
